requirements.txt
pytest-html
pytest
python-dotenv
tkinter
pyttsx3
pygame
psycopg2
Pillow
customtkinter
python-vlc
paho-mqtt
numpy
pyaudio
aubio
playsound
math
subprocess
time
socket